+++ cartridge/failover.lua
@@ -41,6 +41,7 @@ local service_registry = require('cartridge.service-registry')
 local stateboard_client = require('cartridge.stateboard-client')
 local etcd2_client = require('cartridge.etcd2-client')
 local raft_failover = require('cartridge.failover.raft')
+local leader_autoreturn = require('cartridge.failover.leader_autoreturn')
 local argparse = require('cartridge.argparse')
 
 local FailoverError = errors.new_class('FailoverError')
@@ -87,10 +88,6 @@ vars:new('failover_trigger_cnt', 0)
 vars:new('suppress_threshold', math.huge)
 vars:new('suppress_timeout', math.huge)
 vars:new('suppress_fiber', nil)
-
--- autoreturn params
-vars:new('autoreturn_fiber', nil)
-vars:new('autoreturn_delay', math.huge)
 
 
 function _G.__cartridge_failover_get_lsn(timeout)
@@ -647,59 +644,6 @@ local function suppressing_cancel()
     vars.suppress_fiber = nil
 end
 
-local function enable_autoreturn(topology_cfg)
-    while true do
-        fiber.sleep(vars.autoreturn_delay or math.huge)
-        if vars.cache.is_leader then
-            local leaders = topology.get_leaders_order(
-                topology_cfg, vars.replicaset_uuid, nil, {only_enabled = true}
-            )
-            local desired_leader_uuid = leaders[1]
-            if desired_leader_uuid ~= vars.instance_uuid then
-                log.info("Autoreturn: try to return master %s in replicaset %s",
-                    desired_leader_uuid, vars.replicaset_uuid)
-                local client = vars.client
-                if client == nil
-                or client.session == nil
-                or not client.session:is_alive()
-                then
-                    log.error('Autoreturn failed: state provider unavailable')
-                else
-                    local coordinator, err = vars.client.session:get_coordinator()
-                    if coordinator ~= nil and err == nil then
-                        local _, err = package.loaded['cartridge.rpc'].call(
-                            'failover-coordinator',
-                            'appoint_leaders',
-                            {{[vars.replicaset_uuid] = desired_leader_uuid}},
-                            { uri = coordinator.uri }
-                        )
-                        if err ~= nil then
-                            log.error('Autoreturn failed: %s', err.err)
-                        else
-                            log.info('Autoreturn succeded')
-                        end
-                    elseif err ~= nil then
-                        log.error('Autoreturn failed: %s', err.err)
-                    elseif coordinator ~= nil then
-                        log.error('Autoreturn failed: there is no active coordinator')
-                    end
-                end
-            end
-        end
-    end
-end
-
-
-local function autoreturn_cansel()
-    if vars.autoreturn_fiber == nil then
-        return
-    end
-    if vars.autoreturn_fiber:status() ~= 'dead' then
-        vars.autoreturn_fiber:cancel()
-    end
-    vars.autoreturn_fiber = nil
-end
-
 --- Repeatedly fetch new appointments and reconfigure roles.
 --
 -- @function failover_loop
@@ -775,7 +719,7 @@ local function cfg(clusterwide_config, opts)
     end
 
     fencing_cancel()
-    autoreturn_cansel()
+    leader_autoreturn.cansel()
     schedule_clear()
     assert(next(vars.schedule) == nil)
 
@@ -896,10 +840,7 @@ local function cfg(clusterwide_config, opts)
         })
         vars.failover_fiber:name('cartridge.stateful-failover')
 
-        if failover_cfg.master_autoreturn == true then
-            vars.autoreturn_delay = failover_cfg.autoreturn_delay
-            vars.autoreturn_fiber = fiber.new(enable_autoreturn, topology_cfg)
-        end
+        leader_autoreturn.cfg(failover_cfg, topology_cfg)
 
     elseif failover_cfg.mode == 'raft' then
         local ok, err = ApplyConfigError:pcall(raft_failover.check_version)