+++ cartridge/webui/api-failover.lua
@@ -67,7 +67,7 @@ local gql_type_userapi = gql_types.object({
         fencing_enabled = gql_types.boolean.nonNull,
         fencing_timeout = gql_types.float.nonNull,
         fencing_pause = gql_types.float.nonNull,
-        master_autoreturn = gql_types.boolean.nonNull,
+        leader_autoreturn = gql_types.boolean.nonNull,
         autoreturn_delay = gql_types.float.nonNull,
     }
 })
@@ -162,7 +162,7 @@ local function init(graphql)
             fencing_enabled = gql_types.boolean,
             fencing_timeout = gql_types.float,
             fencing_pause = gql_types.float,
-            master_autoreturn = gql_types.boolean,
+            leader_autoreturn = gql_types.boolean,
             autoreturn_delay = gql_types.float,
         },
         kind = gql_type_userapi.nonNull,