+++ webui/src/store/request/queries.graphql.js
@@ -54,7 +54,7 @@ export const getClusterQuery = gql`
         fencing_enabled
         fencing_timeout
         fencing_pause
-        master_autoreturn
+        leader_autoreturn
         autoreturn_delay
         etcd2_params {
           password
@@ -422,7 +422,7 @@ export const changeFailoverMutation = gql`
     $fencing_enabled: Boolean
     $fencing_timeout: Float
     $fencing_pause: Float
-    $master_autoreturn: Boolean
+    $leader_autoreturn: Boolean
     $autoreturn_delay: Float
     $mode: String!
     $state_provider: String
@@ -435,7 +435,7 @@ export const changeFailoverMutation = gql`
         fencing_enabled: $fencing_enabled
         fencing_timeout: $fencing_timeout
         fencing_pause: $fencing_pause
-        master_autoreturn: $master_autoreturn
+        leader_autoreturn: $leader_autoreturn
         autoreturn_delay: $autoreturn_delay
         mode: $mode
         state_provider: $state_provider
@@ -565,7 +565,7 @@ export const getFailoverParams = gql`
         fencing_enabled
         fencing_timeout
         fencing_pause
-        master_autoreturn
+        leader_autoreturn
         autoreturn_delay
         etcd2_params {
           password