+++ CHANGELOG.rst
@@ -57,6 +57,8 @@ Added
 - Compression suggestion, see
   `#1911 <https://github.com/tarantool/tarantool/issues/1911>`_.
 
+- Leader autoreturn feature for stateful failover (`#1949 <https://github.com/tarantool/cartridge/issues/1949>`_).
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Fixed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~